@@ -1,5 +1,7 @@
-from .model import ModelWrapper
-from .mcmc import KurslMCMC
-from .preprocess import Preprocessor
-from .kursl_model import KurSL
+__all__ = ["ModelWrapper", "KurslMCMC", "Preprocessor", "KurSL", "KurslMethod"]
+
 from .kursl_method import KurslMethod
+from .kursl_model import KurSL
+from .mcmc import KurslMCMC
+from .model import ModelWrapper
+from .preprocess import Preprocessor
